fix(narratives): lift metrics charting types out of useBranding.ts
